backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.5
qdrant-client==1.6.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
python-json-logger==2.0.7
